# api/doc/swagger/api.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "https"
  ],
  "swagger": "2.0",
  "info": {
    "title": "standardization-backend API",
    "version": "v1"
  },
  "basePath": "/",
  "paths": {
...
    "/api/standardization/v1/rule/relation/stdfile/{id}": {
      "get": {
        "produces": [
          "application/json"
        ],
        "schemes": [
          "https"
        ],
        "summary": "查询规则关联的标准文件",
        "operationId": "ruleQueryStdFilesByRule",
        "parameters": [
          {
            "minimum": 1,
            "type": "integer",
            "default": 1,
            "name": "offset",
            "in": "query",
            "required": true
          },
          {
            "maximum": 2000,
            "minimum": 0,
            "type": "integer",
            "default": 10,
            "name": "limit",
            "in": "query",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "type": "object",
              "properties": {
                "entries": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "required": [
                      "id",
                      "name",
                      "catalogId",
                      "catalogName"
                    ],
                    "properties": {
                      "catalogId": {
                        "type": "integer"
                      },
                      "catalogName": {
                        "type": "string"
                      },
                      "id": {
                        "type": "integer"
                      },
                      "name": {
                        "type": "string"
                      }
                    }
                  }
                },
                "total_count": {
                  "type": "integer"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/standardization/v1/std-file/{id}": {
      "get": {
        "produces": [
          "application/json"
        ],
        "schemes": [
          "https"
        ],
        "summary": "根据ID查询详情",
        "operationId": "stdfileGetStdFile",
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "type": "object",
              "properties": {
                "actDate": {
                  "type": "string"
                },
                "attachmentType": {
                  "type": "string"
                },
                "attachmentUrl": {
                  "type": "string"
                },
                "catalogId": {
                  "type": "integer"
                },
                "catalogName": {
                  "type": "string"
                },
                "createTime": {
                  "type": "string"
                },
                "createUser": {
                  "type": "string"
                },
                "departmentId": {
                  "type": "string"
                },
                "departmentName": {
                  "type": "string"
                },
                "departmentPathNames": {
                  "type": "string"
                },
                "description": {
                  "type": "string"
                },
                "disableDate": {
                  "type": "string"
                },
                "disableReason": {
                  "type": "string"
                },
                "fileName": {
                  "type": "string"
                },
                "id": {
                  "type": "integer"
                },
                "name": {
                  "type": "string"
                },
                "number": {
                  "type": "string"
                },
                "orgType": {
                  "type": "integer"
                },
                "publishDate": {
                  "type": "string"
                },
                "state": {
                  "type": "string"
                },
                "updateTime": {
                  "type": "string"
                },
                "updateUser": {
                  "type": "string"
                },
                "version": {
                  "type": "integer"
                }
              }
            }
          }
        }
      },
      "put": {
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "produces": [
          "application/json"
        ],
        "schemes": [
          "https"
        ],
        "summary": "根据ID修改标准文件",
        "operationId": "stdfileUpdateStdFile",
        "parameters": [
          {
            "type": "integer",
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "name": "number",
            "in": "formData",
            "allowEmptyValue": true
          },
          {
            "type": "string",
            "name": "name",
            "in": "formData",
            "allowEmptyValue": true
          },
          {
            "type": "integer",
            "name": "catalog_id",
            "in": "formData",
            "allowEmptyValue": true
          },
          {
            "type": "integer",
            "name": "org_type",
            "in": "formData",
            "allowEmptyValue": true
          },
          {
            "type": "string",
            "name": "act_date",
            "in": "formData",
            "allowEmptyValue": true
          },
          {
            "type": "string",
            "name": "publish_date",
            "in": "formData",
            "allowEmptyValue": true
          },
          {
            "type": "string",
            "name": "description",
            "in": "formData",
            "allowEmptyValue": true
          },
          {
            "type": "string",
            "name": "attachment_type",
            "in": "formData",
            "allowEmptyValue": true
          },
          {
            "type": "string",
            "name": "attachment_url",
            "in": "formData",
            "allowEmptyValue": true
          },
          {
            "type": "string",
            "name": "state",
            "in": "formData",
            "allowEmptyValue": true
          },
          {
            "type": "string",
            "name": "department_ids",
            "in": "formData",
            "allowEmptyValue": true
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "type": "object",
              "properties": {
                "actDate": {
                  "type": "string"
                },
                "attachmentType": {
                  "type": "string"
                },
                "attachmentUrl": {
                  "type": "string"
                },
                "catalogId": {
                  "type": "integer"
                },
                "catalogName": {
                  "type": "string"
                },
                "createTime": {
                  "type": "string"
                },
                "createUser": {
                  "type": "string"
                },
                "departmentId": {
                  "type": "string"
                },
                "departmentName": {
                  "type": "string"
                },
                "departmentPathNames": {
                  "type": "string"
                },
                "description": {
                  "type": "string"
                },
                "disableDate": {
                  "type": "string"
                },
                "disableReason": {
                  "type": "string"
                },
                "fileName": {
                  "type": "string"
                },
                "id": {
                  "type": "integer"
                },
                "name": {
                  "type": "string"
                },
                "number": {
                  "type": "string"
                },
                "orgType": {
                  "type": "integer"
                },
                "publishDate": {
                  "type": "string"
                },
                "state": {
                  "type": "string"
                },
                "updateTime": {
                  "type": "string"
                },
                "updateUser": {
                  "type": "string"
                },
                "version": {
                  "type": "integer"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}